Added Node API
