+++ jest.config.js
@@ -4,6 +4,7 @@ module.exports = {
         "!./src/**/*.d.ts",
         "!./src/**/*.stubs.ts",
         "!./src/adapters/*.ts",
+        "!./src/api/*.ts",
         "!./src/cli/main.ts",
         "!./src/converters/editorConfigs/editorSettingsConverters.ts",
         "!./src/converters/lintConfigs/rules/ruleConverters.ts",
