+++ package.json
@@ -58,6 +58,7 @@
             "prettier --write"
         ]
     },
+    "main": "./src/index.js",
     "name": "tslint-to-eslint-config",
     "repository": {
         "type": "git",
