+++ src/cli/main.ts
@@ -1,223 +1,8 @@
 import { EOL } from "os";
 
-import { childProcessExec } from "../adapters/childProcessExec";
-import { fsFileSystem } from "../adapters/fsFileSystem";
-import { globAsync } from "../adapters/globAsync";
-import { nativeImporter } from "../adapters/nativeImporter";
 import { processLogger } from "../adapters/processLogger";
-import { bind } from "../binding";
-import {
-    collectCommentFileNames,
-    CollectCommentFileNamesDependencies,
-} from "../comments/collectCommentFileNames";
-import {
-    ReportCommentResultsDependencies,
-    reportCommentResults,
-} from "../converters/comments/reporting/reportCommentResults";
-import {
-    ConvertEditorConfigDependencies,
-    convertEditorConfig,
-} from "../converters/editorConfigs/convertEditorConfig";
-import {
-    ConvertEditorSettingsDependencies,
-    convertEditorSettings,
-} from "../converters/editorConfigs/convertEditorSettings";
-import { editorSettingsConverters } from "../converters/editorConfigs/editorSettingsConverters";
-import { reportEditorSettingConversionResults } from "../converters/editorConfigs/reporting/reportEditorSettingConversionResults";
-import {
-    ConvertLintConfigDependencies,
-    convertLintConfig,
-} from "../converters/lintConfigs/convertLintConfig";
-import {
-    ReportConversionResultsDependencies,
-    reportConfigConversionResults,
-} from "../converters/lintConfigs/reporting/reportConfigConversionResults";
-import {
-    WriteConversionResultsDependencies,
-    writeConfigConversionResults,
-} from "../converters/lintConfigs/writeConfigConversionResults";
-import {
-    ConvertCommentsDependencies,
-    convertComments,
-} from "../converters/comments/convertComments";
-import {
-    ConvertFileCommentsDependencies,
-    convertFileComments,
-} from "../converters/comments/convertFileComments";
-import {
-    ConvertRulesDependencies,
-    convertRules,
-} from "../converters/lintConfigs/rules/convertRules";
-import { ruleConverters } from "../converters/lintConfigs/rules/ruleConverters";
-import {
-    RetrieveExtendsValuesDependencies,
-    retrieveExtendsValues,
-} from "../converters/lintConfigs/summarization/retrieveExtendsValues";
-import {
-    SummarizePackageRulesDependencies,
-    summarizePackageRules,
-} from "../converters/lintConfigs/summarization/summarizePackageRules";
-import {
-    ChoosePackageManagerDependencies,
-    choosePackageManager,
-} from "../converters/lintConfigs/reporting/packages/choosePackageManager";
-import {
-    LogMissingPackagesDependencies,
-    logMissingPackages,
-} from "../converters/lintConfigs/reporting/packages/logMissingPackages";
-import { runCli, RunCliDependencies } from "./runCli";
-import { ruleMergers } from "../converters/lintConfigs/rules/ruleMergers";
-import { writeEditorConfigConversionResults } from "../converters/lintConfigs/writeEditorConfigConversionResults";
-import { checkPrettierExtension } from "../converters/lintConfigs/summarization/prettier/checkPrettierExtension";
-import { removeExtendsDuplicatedRules } from "../converters/lintConfigs/pruning/removeExtendsDuplicatedRules";
-import {
-    ExtractGlobPathsDependencies,
-    extractGlobPaths,
-} from "../converters/comments/extractGlobPaths";
-import {
-    findEditorConfiguration,
-    FindEditorConfigurationDependencies,
-} from "../input/findEditorConfiguration";
-import { findESLintConfiguration } from "../input/findESLintConfiguration";
-import {
-    findOriginalConfigurations,
-    FindOriginalConfigurationsDependencies,
-} from "../input/findOriginalConfigurations";
-import { findPackagesConfiguration } from "../input/findPackagesConfiguration";
-import { findTSLintConfiguration } from "../input/findTSLintConfiguration";
-import { findTypeScriptConfiguration } from "../input/findTypeScriptConfiguration";
-import { importer, ImporterDependencies } from "../input/importer";
-import { mergeLintConfigurations } from "../input/mergeLintConfigurations";
-
-const convertFileCommentsDependencies: ConvertFileCommentsDependencies = {
-    converters: ruleConverters,
-    fileSystem: fsFileSystem,
-};
-
-const reportCommentResultsDependencies: ReportCommentResultsDependencies = {
-    logger: processLogger,
-};
-
-const convertRulesDependencies: ConvertRulesDependencies = {
-    ruleConverters,
-    ruleMergers,
-};
-
-const convertEditorSettingsDependencies: ConvertEditorSettingsDependencies = {
-    converters: editorSettingsConverters,
-};
-
-const nativeImporterDependencies: ImporterDependencies = {
-    fileSystem: fsFileSystem,
-    getCwd: () => process.cwd(),
-    nativeImporter: nativeImporter,
-};
-
-const boundImporter = bind(importer, nativeImporterDependencies);
-
-const findConfigurationDependencies = {
-    exec: childProcessExec,
-    importer: boundImporter,
-};
-
-const findEditorConfigurationDependencies: FindEditorConfigurationDependencies = {
-    fileSystem: fsFileSystem,
-    importer: boundImporter,
-};
-
-const findOriginalConfigurationsDependencies: FindOriginalConfigurationsDependencies = {
-    findESLintConfiguration: bind(findESLintConfiguration, findConfigurationDependencies),
-    findPackagesConfiguration: bind(findPackagesConfiguration, findConfigurationDependencies),
-    findTypeScriptConfiguration: bind(findTypeScriptConfiguration, findConfigurationDependencies),
-    findTSLintConfiguration: bind(findTSLintConfiguration, findConfigurationDependencies),
-    mergeLintConfigurations,
-};
-
-const collectCommentFileNamesDependencies: CollectCommentFileNamesDependencies = {
-    findTypeScriptConfiguration: bind(findTypeScriptConfiguration, findConfigurationDependencies),
-};
-
-const extractGlobPathsDependencies: ExtractGlobPathsDependencies = {
-    globAsync,
-};
-
-const convertCommentsDependencies: ConvertCommentsDependencies = {
-    collectCommentFileNames: bind(collectCommentFileNames, collectCommentFileNamesDependencies),
-    convertFileComments: bind(convertFileComments, convertFileCommentsDependencies),
-    extractGlobPaths: bind(extractGlobPaths, extractGlobPathsDependencies),
-    reportCommentResults: bind(reportCommentResults, reportCommentResultsDependencies),
-};
-
-const choosePackageManagerDependencies: ChoosePackageManagerDependencies = {
-    fileSystem: fsFileSystem,
-};
-
-const logMissingPackagesDependencies: LogMissingPackagesDependencies = {
-    choosePackageManager: bind(choosePackageManager, choosePackageManagerDependencies),
-    logger: processLogger,
-};
-
-const reportConversionResultsDependencies: ReportConversionResultsDependencies = {
-    logger: processLogger,
-};
-
-const retrieveExtendsValuesDependencies: RetrieveExtendsValuesDependencies = {
-    importer: boundImporter,
-};
-
-const summarizePackageRulesDependencies: SummarizePackageRulesDependencies = {
-    checkPrettierExtension,
-    removeExtendsDuplicatedRules,
-    retrieveExtendsValues: bind(retrieveExtendsValues, retrieveExtendsValuesDependencies),
-};
-
-const writeConversionResultsDependencies: WriteConversionResultsDependencies = {
-    fileSystem: fsFileSystem,
-};
-
-const reportEditorSettingConversionResultsDependencies = {
-    logger: processLogger,
-};
-
-const convertEditorConfigDependencies: ConvertEditorConfigDependencies = {
-    findEditorConfiguration: bind(findEditorConfiguration, findEditorConfigurationDependencies),
-    convertEditorSettings: bind(convertEditorSettings, convertEditorSettingsDependencies),
-    reportEditorSettingConversionResults: bind(
-        reportEditorSettingConversionResults,
-        reportEditorSettingConversionResultsDependencies,
-    ),
-    writeEditorConfigConversionResults: bind(
-        writeEditorConfigConversionResults,
-        writeConversionResultsDependencies,
-    ),
-};
-
-const convertLintConfigDependencies: ConvertLintConfigDependencies = {
-    convertRules: bind(convertRules, convertRulesDependencies),
-    logMissingPackages: bind(logMissingPackages, logMissingPackagesDependencies),
-    reportConfigConversionResults: bind(
-        reportConfigConversionResults,
-        reportConversionResultsDependencies,
-    ),
-    summarizePackageRules: bind(summarizePackageRules, summarizePackageRulesDependencies),
-    writeConfigConversionResults: bind(
-        writeConfigConversionResults,
-        writeConversionResultsDependencies,
-    ),
-};
-
-const runCliDependencies: RunCliDependencies = {
-    converters: [
-        bind(convertLintConfig, convertLintConfigDependencies),
-        bind(convertEditorConfig, convertEditorConfigDependencies),
-        bind(convertComments, convertCommentsDependencies),
-    ],
-    findOriginalConfigurations: bind(
-        findOriginalConfigurations,
-        findOriginalConfigurationsDependencies,
-    ),
-    logger: processLogger,
-};
+import { runCliDependencies } from "../api/dependencies";
+import { runCli } from "../cli/runCli";
 
 export const main = async (argv: string[]) => {
     try {
