+++ src/index.ts
@@ -1,0 +1,2 @@
+export { convertTSLintConfigStandalone as convertTSLintConfig } from "./api/convertTSLintConfigStandalone";
+export * from "./types";
